messages_service/requirements.txt
flask==2.0.1
python-consul==1.1.0
pika==1.2.0
